Lava balancing calculations
I did some balancing calculations for copper, iron and steel to
check that not more stone is used than can be gained by ore cleaning
(even though this is currently not part of the mod).

Lava gain is exponential, but test playing showed that the gain by lava
cooling nicely fits the amount of lava/ores I use in my factory.

Getting the stone is no problem and should be even easier once the
ore cleaning is back in. :)