Dynamics Collector Update
This update adds an automatic range selector. Max size is 1000x1000, and
this scales to the amount of collectors. so 10 collectors means that
each has a range of 100x100, while 20 collectors have a range of 50x50
each.
This is toggeable in the menu, and when you switch it off, you can set
your own range.