Change tech requirements for liquid crafting recipes to match the techs for standard crafting.
Previously, unlocking "Liquid Machine Crafting" would allow all machines to be crafted, even if the appropriate tech was not discovered. Now, both the Liquid Crafting tech (for the mold) and the original tech (for the recipe) are needed before the item can be forged.

Forge recipes for items you begin the game with such as Stone Furnaces are still unlocked by the appropriate Liquid Crafting tech.

Also, Zinc Gear casting is now unlocked with the Zinc Processing tech (an oversight?)