Add "Crafting" to all mold crafting names
With mold crafting recipes now on the technology that provides the base crafting recipe, it looks like the recipes appear twice on the tech screen. Adding "Crafting" to the name is a clue that the recipes are not duplicates, but are Forge recipes.

I considered changing "Crafting" to "Forging", but that should be your call.