Change Forge pipe connections so all liquid icons are visible.
Because the icons previously overlapped, this is needed to find out which liquid goes where, especially for recipes with 3+ liquids.

Note that pipes-to-ground may be *required* for recipes with 5+ liquids. Putting the pipe connections in the corners might have prevented this, but not being able to see the icons kills that idea.