Completely redoing Research System + Lottery
It will now disable technologies all together, and allow every item to
be unlocked.
There are still loads of bugs, but first functionality is in place.
Still lots of work to do though

Lottery:
This is basically the same system that i got in World, but uses science
points to get stuff. Lottery is as realistic as possible. When trying 1
million times, you have a winning percentage of around 0.4%